Refactor languageSwitcher
